gui: now filter items by skills too
Add filtering (not just coloring) of items that don't work for alpha clones. Add ability to control this from a pref pane. This is in contrast to the toolbar button in my earlier commit for the same behavior in the ship browser. The pref should probably cover both.

This is noticeably slow for market groups with a lot of items (like afterburners).