gui: quick hack pass at using current character
For market browser use current fit's active character instead of Alpha clone for filtering. This is nice but doesn't work for ship browser and is flaky if there's no active fit (seems to use the most recently active fit's character, needs investigation).

Also it doesn't refresh automatically when you switch characters yet.

The use case of "no active fit, what ships can my Alpha clone fly" is why I started looking at this in the first place. So, I keep the gold=Omega but now filter for current skills (for items), as separate prefs.

For ship filter I still have the toolbar toggle button and it still just toggles the Omega-only ships. I have some ideas for making it use an actual character and I could easily make it use the pref, but it's not going to be super clean and might not be worth it so I'm keeping it separate while I play with it.

The performance isn't as bad as it was earlier. Needs investigation.